+++ src/demo/App.js
@@ -28,7 +28,9 @@ const demos = {
                 {field: "Tissue Type", type: "nominal", order: "ascending"},
                 {field: "qc_frip", type: "quantitative", order: "descending"}
             ],
-            rowFilter: [ ]
+            rowFilter: [
+                {field: "Tissue Type", type: "nominal", notOneOf: ["None"]}
+            ]
         }
     },
     "H3K27ac Demo (2 Views, Center Tracks)": {
@@ -135,7 +137,17 @@ const demos = {
                     {field: "Species", type: "nominal", position: "right"}
                 ],
                 rowFilter: [
-                    { field: "Cell Type", type: "nominal", contains: "Fibroblast" }
+                    { field: "Cell Type", type: "nominal", notOneOf: [
+                        "Th1", "Spermatid", "ILC1", "Th17", "None", "Monocyte", "Natural Killer Cell", 
+                        "T Lymphocyte", "Erythroid Progenitor Cell", "B Lymphocyte", "liver", "Dendritic Cell", 
+                        "Macrophage", "Myeloid Cell", "Plasmablast", "Th2", "Endothelial Cell", "Epithelium", 
+                        "Melanoma Cell", "Keratinocyte", "Cortex", "Stem cell", "Embryonic Stem Cell", "Adipocyte", 
+                        "Neuroblastoma", "Neuroectoderm", "Neural crest cell", "Glial Cell", "Neural Progenitor Cell", 
+                        "Neuroblastoma patient cells", "Inferior Temporal Lobe Cell", "Substantia Nigra Cell", 
+                        "Hippocampus Middle Cell", "iPSC", "Melanocyte", "Endoderm Cell", "Erythroblast", 
+                        "Lymphoblastoid", "Mesenchymal Stem Cell", "Osteoblast", "Stromal Cell", "Intermediate", 
+                        "Myoblast", "Schwann Cell"
+                    ] }
                 ]
             }
         ]