+++ src/utils/options.js
@@ -103,7 +103,8 @@ const baseSchema = {
         "filterInfo": {
             "type": "object",
             "oneOf":[ 
-                { "required": ["field", "type", "contains"] },
+                { "required": ["field", "type", "notOneOf"] },
+                { "required": ["field", "type", "subtree"] },
                 { "required": ["field", "type", "range"] }
             ],
             "properties": {
@@ -116,8 +117,12 @@ const baseSchema = {
                     "enum": ["nominal", "quantitative", "tree"],
                     "description": "The data type of a field"
                 },
-                "contains": {
-                    "type": ["string", "array"],
+                "notOneOf": {
+                    "type": "array",
+                    "description": "An array of values of a field that should not be included in the filtered data"
+                },
+                "subtree": {
+                    "type": "array",
                     "description": "The substring to search for"
                 },
                 "range": {
@@ -316,6 +321,24 @@ export function addTrackWrapperOptions(options, optionsToAdd, viewId, trackId) {
 }
 
 /**
+ * Get sub-options for a specific track, using its viewId and trackId.
+ * @param {object} options A _processed_ options object to update.
+ * @param {string} key The key of sub-options object.
+ * @param {string} viewId The viewId for the track of interest.
+ * @param {string} trackId The trackId for the track of interest.
+ * @returns {object} The options object for the track, or the default options object.
+ */
+export function getWrapperSubOptions(options, key, viewId, trackId) {
+    if(!options[viewId] || (options[viewId] && !options[viewId][trackId])) {
+        // Update global defaults if there is no track specific options.
+        return options[DEFAULT_OPTIONS_KEY][key];
+    } else {
+        // Update track specific options.
+        return options[viewId][trackId][key];
+    }
+}
+
+/**
  * Update options for a specific track, using its viewId and trackId.
  * @param {object} options A _processed_ options object to update.
  * @param {object} subOptions A sub-options object to replace or insert.
@@ -333,6 +356,7 @@ export function updateWrapperOptions(options, subOptions, key, viewId, trackId, 
             [DEFAULT_OPTIONS_KEY]: {
                 ...options[DEFAULT_OPTIONS_KEY],
                 [key]: isReplace ? subOptions : insertItemToArray(options[DEFAULT_OPTIONS_KEY][key], 0, subOptions)
+                        
             }
         };
     } else {